backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
email-validator
psycopg2-binary
python-dotenv
